backend/app/vector_store.py
# /projects/echovault/backend/app/vector_store.py
import os
from supabase_py import create_client, Client

class VectorStore:
    def __init__(self):
        self.client: Client = self._initialize_client()

    def _initialize_client(self) -> Client:
        """Initializes the Supabase client."""
        url = os.environ.get("SUPABASE_URL")
        key = os.environ.get("SUPABASE_KEY")
        
        if not url or not key:
            raise ValueError("SUPABASE_URL and SUPABASE_KEY must be set in environment.")
            
        return create_client(url, key)

    def search_notes(self, embedding, top_k=5):
        """Searches for similar notes using a vector embedding."""
        # This is a placeholder for the actual RPC call to Supabase
        print(f"Searching for top {top_k} similar notes.")
        return {"status": "success", "results": []}

# Singleton instance
vector_store_client = VectorStore()
