backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
pydantic>=2.5.3
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.7
anyio>=4.2.0
pywebpush>=1.14.0
cryptography>=42.0.0
psycopg2-binary>=2.9.9
google-generativeai>=0.3.2
pgvector>=0.2.4
supabase>=1.0.0
gotrue>=1.0.0